rename internal modules
