+++ commands/info.js
