+++ commands/settings/status/add.js
