+++ commands/settings/status/del.js
