+++ commands/settings/status/view.js
