+++ lib/data-client/index.js
@@ -56,11 +56,11 @@ class DataClient extends Client {
 
     const fromRoot = (path) => join(__dirname, '../..', path)
 
-    this._commandsToLoad = [ fromRoot('command-lib') ]
+    this._commandsToLoad = [ fromRoot('commands') ]
 
-    this._eventsToLoad = [ fromRoot('event-lib') ]
+    this._eventsToLoad = [ fromRoot('events') ]
 
-    this._permissionsToLoad = [ fromRoot('permission-lib') ]
+    this._permissionsToLoad = [ fromRoot('permissions') ]
 
     // load built in events and commands
     this
@@ -181,7 +181,7 @@ class DataClient extends Client {
     for (const fd of files) {
       const filePath = join(path, fd)
       if (
-        (filePath.match(/(?<!\.(?:test|spec))\.[jt]sx?$/)) ||
+        (filePath.match(/(?<!\.(?:test|spec|d))\.[jt]sx?$/)) ||
         (await stat(filePath)).isDirectory()
       ) {
         try {
