+++ permissions/index.d.ts
@@ -1,0 +1,7 @@
+import {
+  Permission
+} from 'eris-boiler'
+
+const admin: Permission
+const owner: Permission
+const vip: Permission
